+++ src/main/java/com/barlinc/unusual_prehistory/UnusualPrehistory2Tab.java
@@ -32,6 +32,7 @@ public class UnusualPrehistory2Tab {
                         output.accept(DROMAEOSAURUS_SPAWN_EGG.get());
                         output.accept(DUNKLEOSTEUS_SPAWN_EGG.get());
                         output.accept(HIBBERTOPTERUS_SPAWN_EGG.get());
+                        output.accept(HYNERPETON_SPAWN_EGG.get());
                         output.accept(JAWLESS_FISH_SPAWN_EGG.get());
                         output.accept(KAPROSUCHUS_SPAWN_EGG.get());
                         output.accept(KENTROSAURUS_SPAWN_EGG.get());
@@ -131,6 +132,7 @@ public class UnusualPrehistory2Tab {
                         output.accept(COELACANTHUS_BUCKET.get());
                         output.accept(DIPLOCAULUS_BUCKET.get());
                         output.accept(DUNKLEOSTEUS_BUCKET.get());
+                        output.accept(HYNERPETON_BUCKET.get());
                         output.accept(JAWLESS_FISH_BUCKET.get());
                         output.accept(KIMMERIDGEBRACHYPTERAESCHNIDIUM_NYMPH_BUCKET.get());
                         output.accept(KIMMERIDGEBRACHYPTERAESCHNIDIUM_BOTTLE.get());