+++ src/main/java/com/barlinc/unusual_prehistory/datagen/client/UP2ItemModelProvider.java
@@ -27,6 +27,7 @@ public class UP2ItemModelProvider extends ItemModelProvider {
     protected void registerModels() {
         this.generatedItem(PRAEPUSA_BUCKET);
         this.generatedItem(BABY_AEGIROCASSIS_BUCKET);
+        this.generatedItem(HYNERPETON_BUCKET);
 
         this.generatedItem(LIVING_OOZE_BUCKET);