+++ src/main/java/com/barlinc/unusual_prehistory/events/ClientModEvents.java
@@ -19,10 +19,7 @@ import com.barlinc.unusual_prehistory.client.models.entity.mob.update_4.lobe_fin
 import com.barlinc.unusual_prehistory.client.models.entity.mob.update_5.*;
 import com.barlinc.unusual_prehistory.client.models.entity.mob.update_5.ambient.DelitzschalaModel;
 import com.barlinc.unusual_prehistory.client.models.entity.mob.update_5.ambient.ZhangsolvaModel;
-import com.barlinc.unusual_prehistory.client.models.entity.mob.update_6.AntarctopeltaModel;
-import com.barlinc.unusual_prehistory.client.models.entity.mob.update_6.CotylorhynchusModel;
-import com.barlinc.unusual_prehistory.client.models.entity.mob.update_6.MammothModel;
-import com.barlinc.unusual_prehistory.client.models.entity.mob.update_6.PrognathodonModel;
+import com.barlinc.unusual_prehistory.client.models.entity.mob.update_6.*;
 import com.barlinc.unusual_prehistory.client.particles.*;
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_1.*;
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_2.OnchopristisRenderer;
@@ -36,10 +33,7 @@ import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_5.GrugRe
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_5.PsilopterusRenderer;
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_5.ambient.DelitzschalaRenderer;
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_5.ambient.ZhangsolvaRenderer;
-import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_6.AntarctopeltaRenderer;
-import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_6.CotylorhynchusRenderer;
-import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_6.MammothRenderer;
-import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_6.PrognathodonRenderer;
+import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_6.*;
 import com.barlinc.unusual_prehistory.registry.*;
 import com.barlinc.unusual_prehistory.screens.TransmogrifierScreen;
 import net.minecraft.client.renderer.BiomeColors;
@@ -144,6 +138,7 @@ public class ClientModEvents {
         // Update 6
         event.registerEntityRenderer(UP2Entities.ANTARCTOPELTA.get(), AntarctopeltaRenderer::new);
         event.registerEntityRenderer(UP2Entities.COTYLORHYNCHUS.get(), CotylorhynchusRenderer::new);
+        event.registerEntityRenderer(UP2Entities.HYNERPETON.get(), HynerpetonRenderer::new);
         event.registerEntityRenderer(UP2Entities.MAMMOTH.get(), MammothRenderer::new);
         event.registerEntityRenderer(UP2Entities.PROGNATHODON.get(), PrognathodonRenderer::new);
     }
@@ -215,6 +210,7 @@ public class ClientModEvents {
         // Update 6
         event.registerLayerDefinition(UP2ModelLayers.ANTARCTOPELTA, AntarctopeltaModel::createBodyLayer);
         event.registerLayerDefinition(UP2ModelLayers.COTYLORHYNCHUS, CotylorhynchusModel::createBodyLayer);
+        event.registerLayerDefinition(UP2ModelLayers.HYNERPETON, HynerpetonModel::createBodyLayer);
         event.registerLayerDefinition(UP2ModelLayers.MAMMOTH, MammothModel::createBodyLayer);
         event.registerLayerDefinition(UP2ModelLayers.PROGNATHODON, PrognathodonModel::createBodyLayer);
     }