+++ src/main/java/com/barlinc/unusual_prehistory/registry/UP2MenuTypes.java
@@ -3,20 +3,15 @@ package com.barlinc.unusual_prehistory.registry;
 import com.barlinc.unusual_prehistory.UnusualPrehistory2;
 import com.barlinc.unusual_prehistory.screens.TransmogrifierMenu;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.world.inventory.MenuType;
-import net.neoforged.neoforge.common.extensions.IMenuTypeExtension;
-import net.neoforged.neoforge.network.IContainerFactory;
 import net.neoforged.neoforge.registries.DeferredHolder;
 import net.neoforged.neoforge.registries.DeferredRegister;
 
 public class UP2MenuTypes {
 
-    public static final DeferredRegister<MenuType<?>> MENUS = DeferredRegister.create(Registries.MENU, UnusualPrehistory2.MOD_ID);
+    public static final DeferredRegister<MenuType<?>> MENU_TYPE = DeferredRegister.create(Registries.MENU, UnusualPrehistory2.MOD_ID);
 
-    public static final DeferredHolder<MenuType<?>, MenuType<TransmogrifierMenu>> TRANSMOGRIFIER = registerMenuType(TransmogrifierMenu::new, "transmogrifier");
+    public static final DeferredHolder<MenuType<?>, MenuType<TransmogrifierMenu>> TRANSMOGRIFIER = MENU_TYPE.register("transmogrifier", () -> new MenuType<>(TransmogrifierMenu::new, FeatureFlags.DEFAULT_FLAGS));
 
-    private static <T extends AbstractContainerMenu> DeferredHolder<MenuType<?>, MenuType<T>> registerMenuType(IContainerFactory<T> factory, String name) {
-        return MENUS.register(name, () -> IMenuTypeExtension.create(factory));
-    }
 }