+++ src/main/java/com/barlinc/unusual_prehistory/registry/tags/UP2ItemTags.java
@@ -145,6 +145,7 @@ public class UP2ItemTags {
     // Update 6
     public static final TagKey<Item> ANTARCTOPELTA_FOOD = modItemTag("antarctopelta_food");
     public static final TagKey<Item> COTYLORHYNCHUS_FOOD = modItemTag("cotylorhynchus_food");
+    public static final TagKey<Item> MAMMOTH_FOOD = modItemTag("mammoth_food");
     public static final TagKey<Item> PROGNATHODON_FOOD = modItemTag("prognathodon_food");
 
     public static final TagKey<Item> SWEET_COTYLORHYNCHUS_FOOD = modItemTag("sweet_cotylorhynchus_food");